ARM: dts: sun6i: Add sun6i-a31s.dtsi
Add a dtsi file for A31s based boards.

Since the  A31s is the same die as the A31 in a different package, this dtsi
simply includes sun6i-a31.dtsi and then overrides the pinctrl compatible to
reflect the different package, everything else is identical.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-include sun6i-a31.dtsi and override the pinctrl compatible, rather then
 copying everything
Changes in v3:
-Do not duplicate the DT path, use label references instead